refactor: class 파일 이동
