+++ src/main/java/step3/utils/StringUtils.java
@@ -1,4 +1,4 @@
-package step3.operation;
+package step3.utils;
 
 public class StringUtils {
