+++ src/test/java/step3/RaceCarTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import step3.operation.Operation;
-import step3.operation.ValidationUtils;
+import step3.utils.ValidationUtils;
 import step3.view.InputView;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -38,20 +38,4 @@ public class RaceCarTest {
         operation.gameStart(numOfAtp, numOfCars);
 
     }
-
-    @DisplayName("자동차 이름 유효성 테스트")
-    @Test
-    public void valid_name() {
-        assertThat(ValidationUtils.validationName("pobi")).isTrue();
-    }
-
-    @DisplayName("자동차 이름 유효성 테스트2")
-    @Test
-    public void invalid_name() {
-        assertThat(ValidationUtils.validationName("kkkkkk")).isFalse();
-        assertThat(ValidationUtils.validationName(" ")).isFalse();
-        assertThat(ValidationUtils.validationName("")).isFalse();
-        assertThat(ValidationUtils.validationName(null)).isFalse();
-    }
-
 }
